+++ ThousandSchnapsen.Simulator/Dto/ActionDto.cs
@@ -3,6 +3,6 @@ namespace ThousandSchnapsen.Simulator.Dto
     public class ActionDto
     {
         public int PlayerId { get; set; }
-        public int CardId { get; set; }
+        public byte CardId { get; set; }
     }
 }
